extract tendermint tests
