+++ consensus/test/test_helpers.go
@@ -1,4 +1,4 @@
-package tendermint
+package test
 
 import (
 	"crypto/ecdsa"
